Agents/content_agent/requirements.txt
FastAPI
pymongo
pinecone
langchain-community
python-dotenv
pymupdf
uvicorn
python-multipart